Improve timeout behavior and display
When a timeout occurs, a modal explaining the matter is displayed,
instead of the weird "try again", "and again" of before.

I also added a spinner gif for the loading buttons.

The german translation is then out-of-date (I just copied the missing
parts from english), and the last four lines of the english one could
probably be improved.